requirements.txt
git+https://github.com/nailbiter/alex_leontiev_toolbox_python@main-poetry

APScheduler==3.6.3
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2020.12.5
charset-normalizer==2.0.9
croniter==1.0.11
cycler==0.11.0
dnspython==2.1.0
fonttools==4.28.5
idna==3.3
iniconfig==1.1.1
kiwisolver==1.3.2
lxml==4.8.0
MarkupSafe==2.0.1
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pymongo==4.0.1
pyparsing==3.0.6
pytest==7.1.1
python-dotenv==0.19.2
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.26.0
schedule==1.1.0
six==1.15.0
soupsieve==2.3.2.post1
tomli==2.0.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.7

ply
json5
more_itertools

gunicorn
uvicorn
fastapi
Flask[async]
